requirements.txt
discord.py
discord-py-slash-commands
PyNaCl
moviepy
tube_dl
requests
pytz
pymongo
dnspython
gTTS